swapped for Mark Blome's implementation now issue with missing symbol when compiling
